Add Safe signer validation and proposal refresh flows
- Reuse block-consistent Safe owner snapshots for signer and signature validation
- Refresh Safe execution routes and proposal nonces before approval
- Add focused coverage for signer routing, confirmations, and simulations
