+++ app/ts/background/safeConfirmationResolver.ts
@@ -6,16 +6,16 @@ import { isSafeEntryWithSafeSigner } from '../types/addressBookTypes.js'
 import { EIP712Message } from '../types/eip721.js'
 import type { SignMessageParams } from '../types/jsonRpc-signing-types.js'
 import type { SafeTransactionSigningRequest } from '../types/safeTypes.js'
-import { METAMASK_ERROR_FAILED_TO_PARSE_REQUEST } from '../utils/constants.js'
 import { checksummedAddress } from '../utils/bigint.js'
 import { getErrorMessage } from '../utils/errors.js'
 import { getPrettySignerName } from '../utils/signerMetadata.js'
 import { modifyObject } from '../utils/typescript.js'
 import { getPendingTransactionsAndMessages, getSafeTransactionStacks, getTabState, getUserAddressBookEntriesForChainIdMorePreciseFirst } from './storageVariables.js'
-import { assertSafeContractStateUnchanged, createSafeTransactionSigningRequest, getSafeContractState, safeTxToTypedDataJson, validateSafeOwnerSignature, validateSafeTransactionForSigning } from '../safe/safeCore.js'
+import { assertSafeContractStateUnchanged, createSafeTransactionSigningRequest, getSafeContractState, safeTxToTypedDataJson, type SafeOwnerValidator, validateSafeTransactionForSigning } from '../safe/safeCore.js'
 import { areSafeExecutionSignerRequestsEqual, prepareSafeExecutionSignerRoute } from '../safe/safeExecutionRouting.js'
 import { reconcileSafeTransactionStack } from '../safe/safeStack.js'
 import type { SafeTx } from '../types/personal-message-definitions.js'
+import { createSafeSignerErrorStatus, type SafeSignerErrorStatus } from './safeSignerErrors.js'
 
 export const SAFE_SIGNER_SELECTION_ERROR_CODE = -32010
 
@@ -23,17 +23,12 @@ type SafeMessageCoSignContext = {
 	readonly safeEntry: SafeEntryWithSafeSigner
 	readonly safeTx: SafeTx
 	readonly safeTxHash: bigint
+	readonly ownerValidator: SafeOwnerValidator
 }
 
 export type RefreshedSafeSignerSelection = {
 	readonly selectedSigner: bigint | undefined
 	readonly verificationError: string | undefined
-}
-
-type SafeSignerErrorStatus = {
-	readonly status: 'SignerError'
-	readonly code: number
-	readonly message: string
 }
 
 export type SafeConfirmationResolution =
@@ -81,20 +76,18 @@ export async function getSafeSignerMismatchApprovalStatus(
 	if (selectedSigner === configuredSafeSigner) return undefined
 	const configuredAddress = checksummedAddress(configuredSafeSigner)
 	if (refreshedSelection?.verificationError !== undefined) {
-		return {
-			status: 'SignerError' as const,
-			code: SAFE_SIGNER_SELECTION_ERROR_CODE,
-			message: `Gnosis Safe signer could not be verified: ${ refreshedSelection.verificationError } Select ${ configuredAddress } in ${ signerName }, then retry.`,
-		}
+		return createSafeSignerErrorStatus(
+			`Gnosis Safe signer could not be verified: ${ refreshedSelection.verificationError } Select ${ configuredAddress } in ${ signerName }, then retry.`,
+			SAFE_SIGNER_SELECTION_ERROR_CODE,
+		)
 	}
 	const selectedAccountDescription = selectedSigner === undefined
 		? 'no account selected'
 		: `${ checksummedAddress(selectedSigner) } selected`
-	return {
-		status: 'SignerError' as const,
-		code: SAFE_SIGNER_SELECTION_ERROR_CODE,
-		message: `Gnosis Safe signer mismatch: this Gnosis Safe is configured to use ${ configuredAddress }, but ${ signerName } currently has ${ selectedAccountDescription }. Select ${ configuredAddress } in ${ signerName }, then retry.`,
-	}
+	return createSafeSignerErrorStatus(
+		`Gnosis Safe signer mismatch: this Gnosis Safe is configured to use ${ configuredAddress }, but ${ signerName } currently has ${ selectedAccountDescription }. Select ${ configuredAddress } in ${ signerName }, then retry.`,
+		SAFE_SIGNER_SELECTION_ERROR_CODE,
+	)
 }
 
 export function getPendingSafeSignerAddress(pending: PendingTransactionOrSignableMessage) {
@@ -131,7 +124,7 @@ async function getSafeMessageCoSignContext(
 	if (reviewedSnapshot.safeAddress !== safeEntry.address || reviewedSnapshot.safeSignerAddress !== safeEntry.safeSignerAddress) {
 		throw new Error('The configured Gnosis Safe signer changed after this co-signing confirmation opened.')
 	}
-	const { safeTxHash, safeState } = await validateSafeTransactionForSigning(
+	const { safeTxHash, safeState, ownerValidator } = await validateSafeTransactionForSigning(
 		ethereum,
 		safeEntry.address,
 		safeEntry.safeSignerAddress,
@@ -140,7 +133,7 @@ async function getSafeMessageCoSignContext(
 	)
 	if (safeTxHash !== reviewedSnapshot.safeTxHash) throw new Error('The Gnosis Safe transaction changed after this co-signing confirmation opened.')
 	assertSafeContractStateUnchanged(reviewedSnapshot.reviewedSafeState, safeState)
-	return { safeEntry, safeTx, safeTxHash }
+	return { safeEntry, safeTx, safeTxHash, ownerValidator }
 }
 
 export async function createSafeMessageCoSignSnapshot(
@@ -180,9 +173,7 @@ export async function validateSafeMessageCoSignature(
 ) {
 	const currentCoSignContext = await getSafeMessageCoSignContext(ethereum, pending)
 	if (currentCoSignContext === undefined) throw new Error('This Gnosis Safe transaction is not eligible for Interceptor co-signing.')
-	const ownerSignature = await validateSafeOwnerSignature(
-		ethereum,
-		currentCoSignContext.safeEntry.address,
+	const ownerSignature = await currentCoSignContext.ownerValidator.validateSignature(
 		currentCoSignContext.safeTxHash,
 		funtypes.String.parse(signerReply),
 		currentCoSignContext.safeEntry.safeSignerAddress,
@@ -190,11 +181,145 @@ export async function validateSafeMessageCoSignature(
 	return ownerSignature.signature
 }
 
-const signerError = (message: string): SafeSignerErrorStatus => ({
-	status: 'SignerError',
-	code: METAMASK_ERROR_FAILED_TO_PARSE_REQUEST,
-	message,
-})
+async function refreshSafeExecutionRoute(
+	ethereum: EthereumClientService,
+	pending: PendingTransactionOrSignableMessage,
+): Promise<PendingTransactionOrSignableMessage | undefined> {
+	if (
+		pending.type !== 'Transaction'
+		|| pending.simulationMode
+		|| pending.safeExecutionOriginalRequestParameters === undefined
+	) return undefined
+	const originalRequest = pending.safeExecutionOriginalRequestParameters
+	const reviewedSafeState = pending.safeExecutionReviewedSafeState
+	if (reviewedSafeState === undefined) throw new Error('Review this Gnosis Safe execution again so its current owner and threshold state can be verified.')
+	const safeAddress = originalRequest.params[0].from
+	if (safeAddress === undefined) throw new Error('The original Gnosis Safe execution request is missing its Gnosis Safe address.')
+	const safeEntry = await getCurrentSafeEntryWithSigner(ethereum, safeAddress)
+	const refreshedRoute = await prepareSafeExecutionSignerRoute(ethereum, originalRequest, safeEntry)
+	if (refreshedRoute === undefined) throw new Error('The configured Gnosis Safe execution route is no longer available.')
+	assertSafeContractStateUnchanged(reviewedSafeState, refreshedRoute.safeState)
+	if (
+		pending.originalRequestParameters.method !== 'eth_sendTransaction'
+		|| !areSafeExecutionSignerRequestsEqual(pending.originalRequestParameters, refreshedRoute.transactionParams)
+	) throw new Error('The Gnosis Safe execution changed after this confirmation opened. Review the refreshed transaction before submitting it.')
+	return modifyObject(pending, {
+		activeAddress: refreshedRoute.executor,
+		originalRequestParameters: refreshedRoute.transactionParams,
+		safeExecutionSignerAddress: refreshedRoute.executor,
+	})
+}
+
+async function getRequiredSafeCoSignContext(
+	ethereum: EthereumClientService,
+	pending: PendingTransactionOrSignableMessage,
+) {
+	if (
+		pending.type !== 'SignableMessage'
+		|| pending.transactionOrMessageCreationStatus !== 'Simulated'
+		|| pending.visualizedPersonalSignRequest.type !== 'SafeTx'
+	) return undefined
+	const context = await getSafeMessageCoSignContext(ethereum, pending)
+	if (context === undefined) throw new Error('This Gnosis Safe transaction is not eligible for Interceptor co-signing.')
+	return context
+}
+
+async function assertSafeProposalSignerReady(
+	ethereum: EthereumClientService,
+	pending: PendingTransactionOrSignableMessage,
+) {
+	if (pending.simulationMode) return undefined
+	if (pending.type === 'Transaction' && pending.safeTransaction !== undefined) {
+		if (pending.safeTransaction.reviewedSafeState === undefined) {
+			throw new Error('Review this Gnosis Safe proposal again so its current owners, threshold, nonce, and signer can be verified.')
+		}
+		await assertReviewedSafeSignerIsStillConfigured(ethereum, pending.safeTransaction)
+	}
+	return undefined
+}
+
+async function getSafeSignerSelectionError(
+	pending: PendingTransactionOrSignableMessage,
+	refreshedSafeSignerSelection?: RefreshedSafeSignerSelection,
+) {
+	if (pending.simulationMode) return undefined
+	const safeSignerAddress = getPendingSafeSignerAddress(pending)
+	if (safeSignerAddress === undefined) return undefined
+	return await getSafeSignerMismatchApprovalStatus(
+		pending.uniqueRequestIdentifier.requestSocket.tabId,
+		safeSignerAddress,
+		refreshedSafeSignerSelection,
+	)
+}
+
+async function refreshSafeProposalNonce(
+	ethereum: EthereumClientService,
+	pending: PendingTransactionOrSignableMessage,
+) {
+	if (pending.simulationMode || pending.type !== 'Transaction' || pending.safeTransaction === undefined) return undefined
+	const currentRequest = pending.safeTransaction
+	const safeState = await getSafeContractState(ethereum, currentRequest.safeAddress)
+	if (currentRequest.reviewedSafeState === undefined) throw new Error('The reviewed Gnosis Safe state is unavailable.')
+	assertSafeContractStateUnchanged(currentRequest.reviewedSafeState, safeState)
+	const storedStackBeforeReconciliation = (await getSafeTransactionStacks()).find((stack) =>
+		stack.chainId === ethereum.getChainId() && stack.safeAddress === currentRequest.safeAddress
+	)
+	const storedStack = storedStackBeforeReconciliation === undefined
+		? undefined
+		: reconcileSafeTransactionStack(storedStackBeforeReconciliation, safeState.nonce)
+	const firstUncommittedNonce = safeState.nonce + BigInt(storedStack?.transactions.length ?? 0)
+	const precedingPendingCount = (await getPendingTransactionsAndMessages()).filter((candidate) =>
+		candidate.type === 'Transaction'
+		&& candidate.safeTransaction?.safeAddress === currentRequest.safeAddress
+		&& candidate.safeTransaction.safeTx.domain.chainId === ethereum.getChainId()
+		&& candidate.safeTransaction.safeTx.message.nonce >= firstUncommittedNonce
+		&& candidate.safeTransaction.safeTx.message.nonce < currentRequest.safeTx.message.nonce
+	).length
+	if (precedingPendingCount > 0) throw new Error('Approve or reject the earlier pending Gnosis Safe proposals before signing this one.')
+	if (currentRequest.safeTx.message.nonce === firstUncommittedNonce) return undefined
+	const executionGasLimit = currentRequest.executionGasLimit ?? (
+		'transactionToSimulate' in pending && pending.transactionToSimulate.success
+			? pending.transactionToSimulate.transaction.gas
+			: undefined
+	)
+	if (executionGasLimit === undefined) throw new Error('The pending Gnosis Safe proposal is missing its execution gas limit.')
+	const refreshedSafeRequest = await createSafeTransactionSigningRequest(
+		ethereum,
+		currentRequest.safeAddress,
+		currentRequest.safeSignerAddress,
+		{
+			to: currentRequest.safeTx.message.to,
+			value: currentRequest.safeTx.message.value,
+			input: currentRequest.safeTx.message.data,
+			gas: executionGasLimit,
+		},
+		firstUncommittedNonce,
+	)
+	return modifyObject(pending, { safeTransaction: refreshedSafeRequest })
+}
+
+function getSafeSignerFacingRequest(
+	pending: PendingTransactionOrSignableMessage,
+	coSignContext: SafeMessageCoSignContext | undefined,
+): SignMessageParams | undefined {
+	if (coSignContext !== undefined) {
+		return {
+			method: 'eth_signTypedData_v4',
+			params: [
+				coSignContext.safeEntry.safeSignerAddress,
+				EIP712Message.parse(safeTxToTypedDataJson(coSignContext.safeTx)),
+			],
+		}
+	}
+	if (pending.type !== 'Transaction' || pending.safeTransaction === undefined) return undefined
+	return {
+		method: 'eth_signTypedData_v4',
+		params: [
+			pending.safeTransaction.safeSignerAddress,
+			EIP712Message.parse(safeTxToTypedDataJson(pending.safeTransaction.safeTx)),
+		],
+	}
+}
 
 export async function resolveSafeConfirmation(
 	ethereum: EthereumClientService,
@@ -206,149 +331,51 @@ export async function resolveSafeConfirmation(
 	let pendingChanged = false
 	if (action !== 'accept') return { status: 'ready', pending, pendingChanged, signerFacingRequest: undefined }
 
-	if (
-		pending.type === 'Transaction'
-		&& !pending.simulationMode
-		&& pending.safeExecutionOriginalRequestParameters !== undefined
-	) {
-		try {
-			const originalRequest = pending.safeExecutionOriginalRequestParameters
-			const reviewedSafeState = pending.safeExecutionReviewedSafeState
-			if (reviewedSafeState === undefined) throw new Error('Review this Gnosis Safe execution again so its current owner and threshold state can be verified.')
-			const safeAddress = originalRequest.params[0].from
-			if (safeAddress === undefined) throw new Error('The original Gnosis Safe execution request is missing its Gnosis Safe address.')
-			const safeEntry = await getCurrentSafeEntryWithSigner(ethereum, safeAddress)
-			const refreshedRoute = await prepareSafeExecutionSignerRoute(ethereum, originalRequest, safeEntry)
-			if (refreshedRoute === undefined) throw new Error('The configured Gnosis Safe execution route is no longer available.')
-			assertSafeContractStateUnchanged(reviewedSafeState, refreshedRoute.safeState)
-			if (
-				pending.originalRequestParameters.method !== 'eth_sendTransaction'
-				|| !areSafeExecutionSignerRequestsEqual(pending.originalRequestParameters, refreshedRoute.transactionParams)
-			) throw new Error('The Gnosis Safe execution changed after this confirmation opened. Review the refreshed transaction before submitting it.')
-			pending = modifyObject(pending, {
-				activeAddress: refreshedRoute.executor,
-				originalRequestParameters: refreshedRoute.transactionParams,
-				safeExecutionSignerAddress: refreshedRoute.executor,
-			})
+	try {
+		const refreshedExecution = await refreshSafeExecutionRoute(ethereum, pending)
+		if (refreshedExecution !== undefined) {
+			pending = refreshedExecution
 			pendingChanged = true
-		} catch (error) {
-			return {
-				status: 'blocked',
-				approvalStatus: signerError(`Gnosis Safe execution could not be prepared: ${ getErrorMessage(error) ?? 'The current Gnosis Safe state could not be validated.' }`),
-			}
+		}
+	} catch (error) {
+		return {
+			status: 'blocked',
+			approvalStatus: createSafeSignerErrorStatus(`Gnosis Safe execution could not be prepared: ${ getErrorMessage(error) ?? 'The current Gnosis Safe state could not be validated.' }`),
 		}
 	}
 
 	let coSignContext: SafeMessageCoSignContext | undefined
-	if (
-		pending.type === 'SignableMessage'
-		&& pending.transactionOrMessageCreationStatus === 'Simulated'
-		&& pending.visualizedPersonalSignRequest.type === 'SafeTx'
-	) {
-		try {
-			coSignContext = await getSafeMessageCoSignContext(ethereum, pending)
-			if (coSignContext === undefined) throw new Error('This Gnosis Safe transaction is not eligible for Interceptor co-signing.')
-		} catch (error) {
-			return {
-				status: 'blocked',
-				approvalStatus: signerError(`Gnosis Safe co-signing request was rejected: ${ getErrorMessage(error) ?? 'The Gnosis Safe transaction could not be validated.' }`),
-			}
+	try {
+		coSignContext = await getRequiredSafeCoSignContext(ethereum, pending)
+	} catch (error) {
+		return {
+			status: 'blocked',
+			approvalStatus: createSafeSignerErrorStatus(`Gnosis Safe co-signing request was rejected: ${ getErrorMessage(error) ?? 'The Gnosis Safe transaction could not be validated.' }`),
 		}
 	}
 
-	if (!pending.simulationMode) {
-		if (pending.type === 'Transaction' && pending.safeTransaction !== undefined) {
-			try {
-				if (pending.safeTransaction.reviewedSafeState === undefined) {
-					throw new Error('Review this Gnosis Safe proposal again so its current owners, threshold, nonce, and signer can be verified.')
-				}
-				await assertReviewedSafeSignerIsStillConfigured(ethereum, pending.safeTransaction)
-			} catch (error) {
-				return {
-					status: 'blocked',
-					approvalStatus: signerError(`Gnosis Safe proposal could not be prepared: ${ getErrorMessage(error) ?? 'The configured signer changed.' }`),
-				}
-			}
+	try {
+		await assertSafeProposalSignerReady(ethereum, pending)
+	} catch (error) {
+		return {
+			status: 'blocked',
+			approvalStatus: createSafeSignerErrorStatus(`Gnosis Safe proposal could not be prepared: ${ getErrorMessage(error) ?? 'The configured signer changed.' }`),
 		}
-		const safeSignerAddress = getPendingSafeSignerAddress(pending)
-		if (safeSignerAddress !== undefined) {
-			const mismatch = await getSafeSignerMismatchApprovalStatus(
-				pending.uniqueRequestIdentifier.requestSocket.tabId,
-				safeSignerAddress,
-				refreshedSafeSignerSelection,
-			)
-			if (mismatch !== undefined) return { status: 'blocked', approvalStatus: mismatch }
+	}
+	const mismatch = await getSafeSignerSelectionError(pending, refreshedSafeSignerSelection)
+	if (mismatch !== undefined) return { status: 'blocked', approvalStatus: mismatch }
+
+	try {
+		const refreshedProposal = await refreshSafeProposalNonce(ethereum, pending)
+		if (refreshedProposal !== undefined) return { status: 'refreshed', pending: refreshedProposal }
+	} catch (error) {
+		return {
+			status: 'blocked',
+			approvalStatus: createSafeSignerErrorStatus(`Gnosis Safe proposal could not be prepared: ${ getErrorMessage(error) ?? 'Failed to refresh the Gnosis Safe nonce.' }`),
 		}
 	}
 
-	if (!pending.simulationMode && pending.type === 'Transaction' && pending.safeTransaction !== undefined) {
-		try {
-			const currentRequest = pending.safeTransaction
-			const safeState = await getSafeContractState(ethereum, currentRequest.safeAddress)
-			if (currentRequest.reviewedSafeState === undefined) throw new Error('The reviewed Gnosis Safe state is unavailable.')
-			assertSafeContractStateUnchanged(currentRequest.reviewedSafeState, safeState)
-			const storedStackBeforeReconciliation = (await getSafeTransactionStacks()).find((stack) =>
-				stack.chainId === ethereum.getChainId() && stack.safeAddress === currentRequest.safeAddress
-			)
-			const storedStack = storedStackBeforeReconciliation === undefined
-				? undefined
-				: reconcileSafeTransactionStack(storedStackBeforeReconciliation, safeState.nonce)
-			const firstUncommittedNonce = safeState.nonce + BigInt(storedStack?.transactions.length ?? 0)
-			const precedingPendingCount = (await getPendingTransactionsAndMessages()).filter((candidate) =>
-				candidate.type === 'Transaction'
-				&& candidate.safeTransaction?.safeAddress === currentRequest.safeAddress
-				&& candidate.safeTransaction.safeTx.domain.chainId === ethereum.getChainId()
-				&& candidate.safeTransaction.safeTx.message.nonce >= firstUncommittedNonce
-				&& candidate.safeTransaction.safeTx.message.nonce < currentRequest.safeTx.message.nonce
-			).length
-			if (precedingPendingCount > 0) throw new Error('Approve or reject the earlier pending Gnosis Safe proposals before signing this one.')
-			if (currentRequest.safeTx.message.nonce !== firstUncommittedNonce) {
-				const executionGasLimit = currentRequest.executionGasLimit ?? (
-					'transactionToSimulate' in pending && pending.transactionToSimulate.success
-						? pending.transactionToSimulate.transaction.gas
-						: undefined
-				)
-				if (executionGasLimit === undefined) throw new Error('The pending Gnosis Safe proposal is missing its execution gas limit.')
-				const refreshedSafeRequest = await createSafeTransactionSigningRequest(
-					ethereum,
-					currentRequest.safeAddress,
-					currentRequest.safeSignerAddress,
-					{
-						to: currentRequest.safeTx.message.to,
-						value: currentRequest.safeTx.message.value,
-						input: currentRequest.safeTx.message.data,
-						gas: executionGasLimit,
-					},
-					firstUncommittedNonce,
-				)
-				return { status: 'refreshed', pending: modifyObject(pending, { safeTransaction: refreshedSafeRequest }) }
-			}
-		} catch (error) {
-			return {
-				status: 'blocked',
-				approvalStatus: signerError(`Gnosis Safe proposal could not be prepared: ${ getErrorMessage(error) ?? 'Failed to refresh the Gnosis Safe nonce.' }`),
-			}
-		}
-	}
-
-	const signerFacingRequest = coSignContext !== undefined
-		? {
-			method: 'eth_signTypedData_v4' as const,
-			params: [
-				coSignContext.safeEntry.safeSignerAddress,
-				EIP712Message.parse(safeTxToTypedDataJson(coSignContext.safeTx)),
-			] as const,
-		}
-		: pending.type === 'Transaction' && pending.safeTransaction !== undefined
-			? {
-				method: 'eth_signTypedData_v4' as const,
-				params: [
-					pending.safeTransaction.safeSignerAddress,
-					EIP712Message.parse(safeTxToTypedDataJson(pending.safeTransaction.safeTx)),
-				] as const,
-			}
-			: undefined
-	return { status: 'ready', pending, pendingChanged, signerFacingRequest }
+	return { status: 'ready', pending, pendingChanged, signerFacingRequest: getSafeSignerFacingRequest(pending, coSignContext) }
 }
 
 type TransactionConfirmationAction = 'accept' | 'reject' | 'noResponse' | 'signerIncluded'
