+++ test/tests/confirmTransactionFlows.test.ts
@@ -1,3 +1,7 @@
+import { beforeEach } from 'bun:test'
+import { resetConfirmTransactionTestState } from './confirmTransactionTestHarness.js'
 import './safeConfirmationFlows.suite.js'
 import './safeStackFlows.suite.js'
 import './terminalReplyDelivery.suite.js'
+
+beforeEach(resetConfirmTransactionTestState)
