+++ test/tests/safeCore.test.ts
@@ -4,7 +4,7 @@ import { AddressBookEntry, getSafeSignerAddresses } from '../../app/ts/types/add
 import { EIP712Message } from '../../app/ts/types/eip721.js'
 import { SafeTx } from '../../app/ts/types/personal-message-definitions.js'
 import { SafeStackExport } from '../../app/ts/types/safeTypes.js'
-import { SAFE_ABI, assertInterceptorSafeTransactionPolicy, assertSafeEoaOwner, assertSafeOwner, createSafeTx, getSafeContractState, getSortedSafeSignatureBytes, normalizeSafeSignature, recoverSafeSignatureOwner, safeTxToTypedDataJson } from '../../app/ts/safe/safeCore.js'
+import { SAFE_ABI, assertInterceptorSafeTransactionPolicy, createSafeOwnerValidator, createSafeTx, getSafeContractSnapshot, getSafeContractState, normalizeSafeSignature, recoverSafeSignatureOwner, safeTxToTypedDataJson } from '../../app/ts/safe/safeCore.js'
 import { completeSafeExecutionWithConfiguredSigner, SAFE_EXECUTION_ABI } from '../../app/ts/safe/safeExecution.js'
 import { getSafeTxHash } from '../../app/ts/utils/eip712.js'
 import { privateKeyToAccount } from '../../app/ts/utils/ethereumPrimitives.js'
@@ -60,19 +60,6 @@ describe('Safe transaction support', () => {
 
 		assert.equal(normalizeSafeSignature(recoveryByteZeroOrOne), signature)
 		assert.equal(await recoverSafeSignatureOwner(BigInt(getSafeTxHash(safeTx)), signature), BigInt(account.address))
-	})
-
-	test('sorts Safe signature bytes by owner address', () => {
-		const firstSignature = `0x${ '11'.repeat(64) }1b`
-		const secondSignature = `0x${ '22'.repeat(64) }1c`
-
-		assert.equal(
-			getSortedSafeSignatureBytes([
-				{ signer: 2n, signature: secondSignature },
-				{ signer: 1n, signature: firstSignature },
-			]),
-			`0x${ firstSignature.slice(2) }${ secondSignature.slice(2) }`,
-		)
 	})
 
 	test('completes a final Safe execution approval with the configured sender instead of another signature request', async () => {
@@ -167,22 +154,21 @@ describe('Safe transaction support', () => {
 		assert.equal(entry.abi, '[]')
 	})
 
-	test('reads Safe metadata and rejects configured non-owners and contract owners', async () => {
+	test('reads Safe metadata and validates configured EOA owners from its snapshot', async () => {
 		const owner = 0x5678n
 		const selectors = {
 			version: encodeFunctionCall(SAFE_ABI, 'VERSION', []).slice(0, 10),
 			nonce: encodeFunctionCall(SAFE_ABI, 'nonce', []).slice(0, 10),
 			owners: encodeFunctionCall(SAFE_ABI, 'getOwners', []).slice(0, 10),
 			threshold: encodeFunctionCall(SAFE_ABI, 'getThreshold', []).slice(0, 10),
-			isOwner: encodeFunctionCall(SAFE_ABI, 'isOwner', ['0x0000000000000000000000000000000000005678']).slice(0, 10),
 		}
-		const createEthereum = (ownerIsValid: boolean) => ({
+		const createEthereum = (contractOwner = false) => ({
 			async getBlockNumber() {
 				return 123n
 			},
-			async getCode(_address: bigint, blockTag: bigint | string) {
+			async getCode(address: bigint, blockTag: bigint | string) {
 				assert.equal(blockTag, 123n)
-				return new Uint8Array([1])
+				return address === 0x1234n || contractOwner ? new Uint8Array([1]) : new Uint8Array()
 			},
 			async call(transaction: { readonly input: Uint8Array }, blockTag: bigint | string) {
 				assert.equal(blockTag, 123n)
@@ -191,24 +177,59 @@ describe('Safe transaction support', () => {
 					case selectors.nonce: return encodeFunctionReturn(SAFE_ABI, 'nonce', [3n])
 					case selectors.owners: return encodeFunctionReturn(SAFE_ABI, 'getOwners', [['0x0000000000000000000000000000000000005678']])
 					case selectors.threshold: return encodeFunctionReturn(SAFE_ABI, 'getThreshold', [2n])
-					case selectors.isOwner: return encodeFunctionReturn(SAFE_ABI, 'isOwner', [ownerIsValid])
 					default: throw new Error('Unexpected Safe call')
 				}
 			},
 		})
 
-		assert.deepEqual(await getSafeContractState(createEthereum(true), 0x1234n), {
+		assert.deepEqual(await getSafeContractState(createEthereum(), 0x1234n), {
 			version: '1.4.1',
 			nonce: 3n,
 			owners: [owner],
 			threshold: 2n,
 		})
-		await assert.doesNotReject(assertSafeOwner(createEthereum(true), 0x1234n, owner))
-		await assert.rejects(assertSafeOwner(createEthereum(false), 0x1234n, owner), /is not an owner of Gnosis Safe/u)
-		await assert.doesNotReject(assertSafeEoaOwner({
-			...createEthereum(true),
-			async getCode() { return new Uint8Array() },
-		}, 0x1234n, owner))
-		await assert.rejects(assertSafeEoaOwner(createEthereum(true), 0x1234n, owner), /supports EOA owners only/u)
+		const snapshot = await getSafeContractSnapshot(createEthereum(), 0x1234n)
+		await assert.doesNotReject(createSafeOwnerValidator(createEthereum(), 0x1234n, snapshot).assertEoaOwner(owner))
+		await assert.rejects(createSafeOwnerValidator(createEthereum(), 0x1234n, snapshot).assertEoaOwner(0x9abcn), /is not an owner of Gnosis Safe/u)
+		await assert.rejects(createSafeOwnerValidator(createEthereum(true), 0x1234n, snapshot).assertEoaOwner(owner), /supports EOA owners only/u)
+	})
+
+	test('reuses the Safe snapshot and owner-code result across repeated validations', async () => {
+		let blockNumberReads = 0
+		let ownerCodeReads = 0
+		const blockTags: (bigint | string)[] = []
+		const baseEthereum = {
+			async getBlockNumber() {
+				blockNumberReads += 1
+				return 122n + BigInt(blockNumberReads)
+			},
+			async getCode(address: bigint, blockTag: bigint | string) {
+				blockTags.push(blockTag)
+				if (address !== 0x1234n) ownerCodeReads += 1
+				return address === 0x1234n ? new Uint8Array([1]) : new Uint8Array()
+			},
+			async call(transaction: { readonly input: Uint8Array }, blockTag: bigint | string) {
+				blockTags.push(blockTag)
+				switch (bytesToHex(transaction.input).slice(0, 10)) {
+					case encodeFunctionCall(SAFE_ABI, 'VERSION', []).slice(0, 10): return encodeFunctionReturn(SAFE_ABI, 'VERSION', ['1.4.1'])
+					case encodeFunctionCall(SAFE_ABI, 'nonce', []).slice(0, 10): return encodeFunctionReturn(SAFE_ABI, 'nonce', [3n])
+					case encodeFunctionCall(SAFE_ABI, 'getOwners', []).slice(0, 10): return encodeFunctionReturn(SAFE_ABI, 'getOwners', [['0x0000000000000000000000000000000000005678']])
+					case encodeFunctionCall(SAFE_ABI, 'getThreshold', []).slice(0, 10): return encodeFunctionReturn(SAFE_ABI, 'getThreshold', [2n])
+					default: throw new Error('Unexpected Safe call')
+				}
+			},
+		}
+
+		const snapshot = await getSafeContractSnapshot(baseEthereum, 0x1234n)
+		const ownerValidator = createSafeOwnerValidator(baseEthereum, 0x1234n, snapshot)
+		await Promise.all([
+			ownerValidator.assertEoaOwner(0x5678n),
+			ownerValidator.assertEoaOwner(0x5678n),
+		])
+
+		assert.equal(blockNumberReads, 1)
+		assert.equal(ownerCodeReads, 1)
+		assert.equal(snapshot.blockNumber, 123n)
+		assert.equal(blockTags.every((blockTag) => blockTag === snapshot.blockNumber), true)
 	})
 })
